Add files via upload
This is probably the simplest and easiest game you will ever play, so sorry to people who were looking for more of a challenge. Just click anywhere within bounds (white rectangle) to place the football and your character will spawn at a random location. Use the arrow keys to move to the ball and then to the end zone (red squares). Avoid the defenders. Or you can press 'a' and your character will use Astar to get to the ball then the end zone.